Feat: utils 객체에 type string을 가져올 수 있는 기능 추가
해당 기능을 추가후 utils 객체를 Common.js라는 파일로 따로 분리시켰다.
이유는 후에 추가할 fetch 모듈이 type string을 가져오는 기능을 사용하는데
fetch 모듈이 index.js에 import 되면 의존성 사이클이 생겨 문제가 생긴다.
